+++ Dockerfile
@@ -1,10 +1,13 @@
 FROM python:3.12
 
+# TODO assert C.UTF8 and PYTHONUNBUFFERED are set correctly
+
 WORKDIR /srv
 RUN --mount=type=cache,target=/root/.cache pip install --upgrade pip-tools wheel
 COPY requirements.txt .
 RUN --mount=type=cache,target=/root/.cache pip-sync
 
+# TODO multi-stage for faster development builds?
 RUN mkdir -p static
 
 COPY . .
