+++ requirements.txt
@@ -2,22 +2,59 @@
 # This file is autogenerated by pip-compile with Python 3.11
 # by the following command:
 #
-#    pip-compile --no-emit-index-url
+#    pip-compile
 #
+anyio==4.3.0
+    # via
+    #   httpx
+    #   jupyter-server
+appnope==0.1.4
+    # via ipykernel
+argon2-cffi==23.1.0
+    # via jupyter-server
+argon2-cffi-bindings==21.2.0
+    # via argon2-cffi
+arrow==1.3.0
+    # via isoduration
 asgiref==3.7.2
     # via django
+asttokens==2.4.1
+    # via stack-data
+async-lru==2.0.4
+    # via jupyterlab
+attrs==23.2.0
+    # via
+    #   jsonschema
+    #   referencing
+babel==2.14.0
+    # via jupyterlab-server
+beautifulsoup4==4.12.3
+    # via nbconvert
+bleach==6.1.0
+    # via nbconvert
 certifi==2023.5.7
-    # via requests
+    # via
+    #   httpcore
+    #   httpx
+    #   requests
 cffi==1.16.0
-    # via cryptography
+    # via
+    #   argon2-cffi-bindings
+    #   cryptography
 cfgv==3.3.1
     # via pre-commit
 charset-normalizer==3.1.0
     # via requests
+comm==0.2.1
+    # via ipykernel
 cryptography==41.0.5
-    # via
-    #   -r requirements.in
-    #   secretstorage
+    # via -r requirements.in
+debugpy==1.8.1
+    # via ipykernel
+decorator==5.1.1
+    # via ipython
+defusedxml==0.7.1
+    # via nbconvert
 distlib==0.3.6
     # via virtualenv
 django==4.2.1
@@ -44,28 +81,107 @@ djangorestframework-stubs[compatible-mypy]==3.14.0
     # via -r requirements.in
 docutils==0.20.1
     # via readme-renderer
+executing==2.0.1
+    # via stack-data
+fastjsonschema==2.19.1
+    # via nbformat
 filelock==3.12.0
     # via virtualenv
+fqdn==1.5.1
+    # via jsonschema
+h11==0.14.0
+    # via httpcore
+httpcore==1.0.4
+    # via httpx
+httpx==0.27.0
+    # via jupyterlab
 identify==2.5.24
     # via pre-commit
 idna==3.4
-    # via requests
+    # via
+    #   anyio
+    #   httpx
+    #   jsonschema
+    #   requests
 importlib-metadata==7.0.1
     # via
     #   keyring
     #   twine
+ipykernel==6.29.3
+    # via jupyterlab
+ipython==8.22.2
+    # via ipykernel
+isoduration==20.11.0
+    # via jsonschema
 jaraco-classes==3.3.0
     # via keyring
-jeepney==0.8.0
+jedi==0.19.1
+    # via ipython
+jinja2==3.1.3
     # via
-    #   keyring
-    #   secretstorage
+    #   jupyter-server
+    #   jupyterlab
+    #   jupyterlab-server
+    #   nbconvert
+json5==0.9.22
+    # via jupyterlab-server
+jsonpointer==2.4
+    # via jsonschema
+jsonschema[format-nongpl]==4.21.1
+    # via
+    #   jupyter-events
+    #   jupyterlab-server
+    #   nbformat
+jsonschema-specifications==2023.12.1
+    # via jsonschema
+jupyter-client==8.6.0
+    # via
+    #   ipykernel
+    #   jupyter-server
+    #   nbclient
+jupyter-core==5.7.1
+    # via
+    #   ipykernel
+    #   jupyter-client
+    #   jupyter-server
+    #   jupyterlab
+    #   nbclient
+    #   nbconvert
+    #   nbformat
+jupyter-events==0.9.0
+    # via jupyter-server
+jupyter-lsp==2.2.4
+    # via jupyterlab
+jupyter-server==2.13.0
+    # via
+    #   jupyter-lsp
+    #   jupyterlab
+    #   jupyterlab-server
+    #   notebook-shim
+jupyter-server-terminals==0.5.2
+    # via jupyter-server
+jupyterlab==4.1.4
+    # via -r requirements.in
+jupyterlab-pygments==0.3.0
+    # via nbconvert
+jupyterlab-server==2.25.3
+    # via jupyterlab
 keyring==24.3.0
     # via twine
 markdown-it-py==3.0.0
     # via rich
+markupsafe==2.1.5
+    # via
+    #   jinja2
+    #   nbconvert
+matplotlib-inline==0.1.6
+    # via
+    #   ipykernel
+    #   ipython
 mdurl==0.1.2
     # via markdown-it-py
+mistune==3.0.2
+    # via nbconvert
 more-itertools==10.2.0
     # via jaraco-classes
 mypy==1.2.0
@@ -75,48 +191,166 @@ mypy==1.2.0
     #   djangorestframework-stubs
 mypy-extensions==1.0.0
     # via mypy
+nbclient==0.9.0
+    # via nbconvert
+nbconvert==7.16.2
+    # via jupyter-server
+nbformat==5.9.2
+    # via
+    #   jupyter-server
+    #   nbclient
+    #   nbconvert
+nest-asyncio==1.6.0
+    # via ipykernel
 nh3==0.2.15
     # via readme-renderer
 nodeenv==1.8.0
     # via pre-commit
+notebook-shim==0.2.4
+    # via jupyterlab
+overrides==7.7.0
+    # via jupyter-server
+packaging==24.0
+    # via
+    #   ipykernel
+    #   jupyter-server
+    #   jupyterlab
+    #   jupyterlab-server
+    #   nbconvert
+pandocfilters==1.5.1
+    # via nbconvert
+parso==0.8.3
+    # via jedi
+pexpect==4.9.0
+    # via ipython
 pkginfo==1.9.6
     # via twine
 platformdirs==3.5.1
-    # via virtualenv
+    # via
+    #   jupyter-core
+    #   virtualenv
 pre-commit==3.3.2
     # via -r requirements.in
+prometheus-client==0.20.0
+    # via jupyter-server
+prompt-toolkit==3.0.43
+    # via ipython
+psutil==5.9.8
+    # via ipykernel
+ptyprocess==0.7.0
+    # via
+    #   pexpect
+    #   terminado
+pure-eval==0.2.2
+    # via stack-data
 pycparser==2.21
     # via cffi
 pygments==2.17.2
     # via
+    #   ipython
+    #   nbconvert
     #   readme-renderer
     #   rich
 pyjwt==2.8.0
     # via -r requirements.in
+python-dateutil==2.9.0.post0
+    # via
+    #   arrow
+    #   jupyter-client
+python-json-logger==2.0.7
+    # via jupyter-events
 pytz==2024.1
     # via djangorestframework
 pyyaml==6.0.1
-    # via pre-commit
+    # via
+    #   jupyter-events
+    #   pre-commit
+pyzmq==25.1.2
+    # via
+    #   ipykernel
+    #   jupyter-client
+    #   jupyter-server
 readme-renderer==42.0
     # via twine
+referencing==0.33.0
+    # via
+    #   jsonschema
+    #   jsonschema-specifications
+    #   jupyter-events
 requests==2.31.0
     # via
     #   -r requirements.in
     #   djangorestframework-stubs
+    #   jupyterlab-server
     #   requests-toolbelt
     #   twine
 requests-toolbelt==1.0.0
     # via twine
+rfc3339-validator==0.1.4
+    # via
+    #   jsonschema
+    #   jupyter-events
 rfc3986==2.0.0
     # via twine
+rfc3986-validator==0.1.1
+    # via
+    #   jsonschema
+    #   jupyter-events
 rich==13.7.0
     # via twine
-secretstorage==3.3.3
-    # via keyring
+rpds-py==0.18.0
+    # via
+    #   jsonschema
+    #   referencing
+send2trash==1.8.2
+    # via jupyter-server
+six==1.16.0
+    # via
+    #   asttokens
+    #   bleach
+    #   python-dateutil
+    #   rfc3339-validator
+sniffio==1.3.1
+    # via
+    #   anyio
+    #   httpx
+soupsieve==2.5
+    # via beautifulsoup4
 sqlparse==0.4.4
     # via django
+stack-data==0.6.3
+    # via ipython
+terminado==0.18.0
+    # via
+    #   jupyter-server
+    #   jupyter-server-terminals
+tinycss2==1.2.1
+    # via nbconvert
+tornado==6.4
+    # via
+    #   ipykernel
+    #   jupyter-client
+    #   jupyter-server
+    #   jupyterlab
+    #   terminado
+traitlets==5.14.1
+    # via
+    #   comm
+    #   ipykernel
+    #   ipython
+    #   jupyter-client
+    #   jupyter-core
+    #   jupyter-events
+    #   jupyter-server
+    #   jupyterlab
+    #   matplotlib-inline
+    #   nbclient
+    #   nbconvert
+    #   nbformat
 twine==4.0.2
     # via -r requirements.in
+types-python-dateutil==2.8.19.20240311
+    # via arrow
 types-pytz==2023.3.0.0
     # via django-stubs
 types-pyyaml==6.0.12.10
@@ -135,12 +369,24 @@ typing-extensions==4.6.3
     #   django-stubs-ext
     #   djangorestframework-stubs
     #   mypy
+uri-template==1.3.0
+    # via jsonschema
 urllib3==2.0.3
     # via
     #   requests
     #   twine
 virtualenv==20.23.0
     # via pre-commit
+wcwidth==0.2.13
+    # via prompt-toolkit
+webcolors==1.13
+    # via jsonschema
+webencodings==0.5.1
+    # via
+    #   bleach
+    #   tinycss2
+websocket-client==1.7.0
+    # via jupyter-server
 zipp==3.17.0
     # via importlib-metadata
